merge

+++ src/agent/agent_impl.cc
@@ -16,7 +16,7 @@ extern std::string FLAGS_master_addr;
 extern std::string FLAGS_agent_port;
 extern std::string FLAGS_agent_work_dir;
 extern double FLAGS_cpu_num;
-extern int FLAGS_mem_gbytes;
+extern int64_t FLAGS_mem_gbytes;
 
 namespace galaxy {
